LayoutTests/fast/text/accessibility-bold-system-font-expected.txt

This test makes sure that text using the accessibility bold functionality has the correct weight.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(target).getPropertyValue('font-weight') is "400"
PASS successfullyParsed is true

TEST COMPLETE
Hello
